requirements.txt
mcp
google-generativeai
pillow